[Dropdown] Allow ignoring of case when getting an item
Currently get.item compares labels and values case sensitively. This can cause duplicates when allowAdditions is true. To prevent these duplicates get.item should be capable of comparing labels and values while ignoring case.

However the current behaviour should be preserved by default.

To achieve this a new setting is added called ignoresCase, which when true will disregard the casing of labels and string values when performing the comparison.